Re wrote parts of the program that used random ints, to random floats. Remember kids always read the asignment in full before starting.

+++ worker.c
@@ -3,13 +3,15 @@
 #include <sys/ipc.h>
 #include <sys/msg.h>
 #include <stdio.h>
+#include <string.h>
+#include <stdlib.h>
 
 //parameters workerID nBuffers sleepTie msgID shmID semID
 int main(int argc, char* argv[]){
 	struct message newM;
 	newM.type = HELLO;
 	newM.workerID = atoi(argv[1]);
-	newM.sleepTime = atoi(argv[3]);
+	newM.sleepTime = atof(argv[3]);
 
 	int msgID = atoi(argv[4]);
 	if(msgsnd(msgID, &newM, sizeof(struct message) - sizeof(long int), IPC_NOWAIT) < 0){
